requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
python-decouple==3.8
redis==5.0.4
sqlparse==0.5.0
tzdata==2024.1
